Enable database logging
in Odoo >= 8.0 --log_db <dbname> allows logging in database.
The logs are readable in Settings > Technical > Database Structure > Logging

This commit enables the setting by default with a in database logging level of
warning to avoid having too much data in there. I found this helpful to be able
to check the presence of warnings and errors from the user interface rather
than having to connect to a shell.