bump reportlab to 3.1.44
With reportlab 2.5 I get unicode-related crashes when trying to print
ean13 barcodes This applies for example to the report of the picking
where a product has an ean13 code).

The latest version of reportlab as of today is 3.2, but 3.1.44 is the
one from the odoo 8.0 requirements.txt.

Sidenote: I thought that the anybox recipe fetched by default the
versions found in requirements.txt, but apparently it is not the case.