Do not use index http://pypi.camptocamp.net/pypi
As discussed on our mailing-list, we are not sure that we still need our
local index.  We will no longer use it and see.  If we have issues with
the pypi index, we'll consider to enable it again.

One of the issue is that Travis uses unnecessary bandwidth for the eggs.